py/requirements.txt
Pillow==11.1.0
qrcode==8.0
